+++ msaduana/src/main/java/com/aduanas/msaduana/repository/VehiculoRepository.java
@@ -9,5 +9,6 @@ import java.util.Optional;
 @Repository
 public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
     Optional<Vehiculo> findByPatente(String patente);
+    Optional<Vehiculo> findByRutDuenio(String rutDuenio);
     boolean existsByPatente(String patente);
 }