+++ msaduana/src/main/java/com/aduanas/msaduana/service/VehiculoService.java
@@ -2,7 +2,6 @@ package com.aduanas.msaduana.service;
 
 import com.aduanas.msaduana.dto.VehiculoRequestDTO;
 import com.aduanas.msaduana.dto.VehiculoResponseDTO;
-import com.aduanas.msaduana.model.Vehiculo;
 
 import java.util.List;
 
@@ -10,6 +9,7 @@ public interface VehiculoService {
     VehiculoResponseDTO crearVehiculo(VehiculoRequestDTO request);
     VehiculoResponseDTO buscarVehiculo(Long id);
     VehiculoResponseDTO buscarPorPatente(String patente);
+    VehiculoResponseDTO buscarPorRut(String rutDuenio);
     List<VehiculoResponseDTO> listarVehiculos();
     VehiculoResponseDTO autorizarPaso(Long id, boolean autorizar);
 }