+++ msviajeros/src/main/java/com/aduanas/msviajeros/dto/ValidacionDTO.java
@@ -1,0 +1,7 @@
+package com.aduanas.msviajeros.dto;
+
+public class ValidacionDTO {
+        private String estado;
+        public String getEstado() { return estado; }
+        public void setEstado(String estado) { this.estado = estado; }
+}