+++ msviajeros/src/main/java/com/aduanas/msviajeros/model/Viajero.java
@@ -1,11 +1,14 @@
 package com.aduanas.msviajeros.model;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,31 +23,34 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Viajero {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+        @Id
+        @GeneratedValue(strategy = GenerationType.IDENTITY)
+        private Long id;
 
-    private String rut;
+        private String rut;
 
-    private String nombres;
+        private String nombres;
 
-    private String apellidos;
+        private String apellidos;
 
-    private String documento;
+        private String documento;
 
-    private String nacionalidad;
+        private String nacionalidad;
 
-    private String vehiculo;
+        private String vehiculo;
 
-    private String patente;
+        private String patente;
 
-    private String origen;
+        private String origen;
 
-    private String destino;
+        private String destino;
 
-    private String estado;
+        private String estado;
 
-    private String riesgo;
+        private String riesgo;
 
-    private LocalDateTime fechaIngreso;
+        private LocalDateTime fechaIngreso;
+
+        @OneToMany(mappedBy = "viajero", cascade = CascadeType.ALL, orphanRemoval = true)
+        private List<Menor> menores;
 }