+++ msviajeros/src/main/java/com/aduanas/msviajeros/repository/ViajeroRepository.java
@@ -1,5 +1,6 @@
 package com.aduanas.msviajeros.repository;
 
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.aduanas.msviajeros.model.Viajero;
@@ -9,4 +10,6 @@ public interface ViajeroRepository extends JpaRepository<Viajero, Long> {
     Optional<Viajero> findByRut(String rut);
 
     boolean existsByRut(String rut);
+
+    List<Viajero> findByEstado(String estado);
 }